Add files via upload
In this Project I was given a dataset in which I was asked to create insights on the basis of client named "Social Buzz'. There were several question that the client asked like, for the growth of the company and which type content was consumed the most and many other.